organize commands into subfolders.
