+++ src/commands/fun/affirmation.ts
@@ -1,7 +1,7 @@
 import { Command } from '@knighthacks/dispatch';
 import axios from 'axios';
 import { Message } from 'discord.js';
-import { singleButtonRow } from '../util/button';
+import { singleButtonRow } from '../../util/button';
 
 const url = 'https://www.affirmations.dev/';
 const row = singleButtonRow({
