+++ src/commands/fun/dice.ts
