+++ src/commands/fun/eightball.ts
