+++ src/commands/util/testCommand.ts
@@ -1,5 +1,5 @@
 import { Command, inChannelNames } from '@knighthacks/dispatch';
-import { Channels } from '../channels';
+import { Channels } from '../../channels';
 
 const command: Command = {
   name: 'test',
